modify response model
