+++ OffsiteTest/API/Core/APIList.swift
@@ -1,0 +1,13 @@
+//
+//  APIList.swift
+//  OffsiteTest
+//
+//  Created by Thomas Lam on 21/4/2018.
+//  Copyright © 2018 Thomas Lam. All rights reserved.
+//
+
+public struct APIList {
+    static let topNFreeApp = "https://itunes.apple.com/hk/rss/topfreeapplications/limit=%d/json"
+    static let topNGrossingApp = "https://itunes.apple.com/hk/rss/topgrossingapplications/limit=%d/json"
+    static let appDetail = "https://itunes.apple.com/hk/lookup?id=%@"
+}
