--- OffsiteTest/API/Response/Listing/Feed.swift
@@ -1,49 +1,0 @@
-/* 
-Copyright (c) 2018 Swift Models Generated from JSON powered by http://www.json4swift.com
-
-Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-
-For support, please feel free to contact me at https://www.linkedin.com/in/syedabsar
-
-*/
-
-import Foundation
-struct Feed : Codable {
-	let author : Author?
-	let entry : [Entry]?
-	let updated : Updated?
-	let rights : Rights?
-	let title : Title?
-	let icon : Icon?
-	let link : [Link]?
-	let id : Id?
-
-	enum CodingKeys: String, CodingKey {
-
-		case author
-		case entry = "entry"
-		case updated
-		case rights
-		case title
-		case icon
-		case link = "link"
-		case id
-	}
-
-	init(from decoder: Decoder) throws {
-		let values = try decoder.container(keyedBy: CodingKeys.self)
-		author = try Author(from: decoder)
-		entry = try values.decodeIfPresent([Entry].self, forKey: .entry)
-		updated = try Updated(from: decoder)
-		rights = try Rights(from: decoder)
-		title = try Title(from: decoder)
-		icon = try Icon(from: decoder)
-		link = try values.decodeIfPresent([Link].self, forKey: .link)
-		id = try Id(from: decoder)
-	}
-
-}
