--- OffsiteTest/ViewController.swift
@@ -1,25 +1,0 @@
-//
-//  ViewController.swift
-//  OffsiteTest
-//
-//  Created by Thomas Lam on 21/4/2018.
-//  Copyright © 2018 Thomas Lam. All rights reserved.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view, typically from a nib.
-    }
-
-    override func didReceiveMemoryWarning() {
-        super.didReceiveMemoryWarning()
-        // Dispose of any resources that can be recreated.
-    }
-
-
-}
-
